PeacePath_Code/requirements.txt
streamlit==1.30.0
streamlit-option-menu==0.3.6
pandas==2.0.0
opencv-python==4.10.0.84
numpy==1.26.4
tensorflow==2.15.0
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.12.2
SpeechRecognition==3.10.0
deepface==0.0.81
google-generativeai==0.3.2
